+++ resources/views/laraknife/icon/edit.blade.php
@@ -1,1 +1,0 @@
-<a href="{{$ref}}"><img src="resources/icons/edit.png" alt="Ändern des Datensatzes"></a>