+++ resources/views/note/edit.blade.php
@@ -13,6 +13,10 @@
                     value="{{ $context->valueOf('title') }}" width2="10" />
                 <x-laraknife.forms.text position="alone" name="body" label="Body"
                     value="{{ $context->valueOf('body') }}" width2="10" rows="10" />
+                <div class="row lkn-empty-line-above">
+                    <x-laraknife.buttons.store width1="2" width2="4" />
+                    <x-laraknife.buttons.cancel width1="2" width2="4" />
+                </div>
             </x-laraknife.layout.nav-tabs>
         </x-laraknife.panels.standard>
     </form>